pyproject.toml
[tool.poetry]
name = "gameassistant"
version = "0.1.0"
description = ""
authors = ["pacificbelt30 <57101176+pacificbelt30@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyautogui = "^0.9.54"
pillow = "^10.4.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
